Refactoring native asset min lovelace calculation into 2 parts so calculations can be done on token bundles before tx output creation
